Time: 0 ms (100%), Space: 14 MB (56.42%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.cpp
@@ -4,7 +4,7 @@ public:
     vector<int> singleNumber(vector<int>& nums) {
         int n = nums.size();
 
-        int xorr = 0;
+        long xorr = 0;
 
         for (int i = 0; i < n; i++) {
             xorr = xorr ^ nums[i];
